component for automatic login/logout-button
